improve many controllers; protect few routes

+++ controllers/PaymentController.js
@@ -1,52 +1,61 @@
-const { Payment } = require("../db/models")
+const { Payment, OrderHeader } = require("../db/models")
 
 module.exports = {
-    // create new Payment
     create: async (req,res) => {
-        const payment = req.body;
-        await Payment.create(payment);
-        res.json(payment);
+        if (!req?.body?.name)
+            return res.status(400).json({ 'message': 'Name parameter not specified.' });
+        const payment = await Payment.findOne({ where: { name: req.body.name } });
+        if(payment)
+            return res.status(204).json({ 'message': 'Payment with same Name already exists.' });
+        const newPayment = req.body;
+        await Payment.create(newPayment);
+        res.json(newPayment);
     },
-    // update Payment
+    
     update: async (req,res) => {
-        const id = req.params.id;
-        const updated = await Payment.update(
-            { 
-                name: req.body.name
-            }, 
-            {
-            where: {
-                id: id
-            }
-            });
-
+        if (!req?.body?.name)
+            return res.status(400).json({ 'message': 'Name parameter not specified.' });
+        const payment = await Payment.findOne({ where: { id: req.params.id } });
+        if(!payment)
+            return res.status(204).json({ 'message': `No Payment matching Id ${req.params.id} has been found.` });
+        await Payment.update(
+            { name: req.body.name }, 
+            { where: { id: req.params.id } }
+        );
         res.json("Updated successfully.");
     },
-    // delete Payment
+    
     delete: async (req,res) => {
-        const id = req.params.id;
+        const payment = await Payment.findOne({ where: { id: req.params.id } });
+        if(!payment)
+            return res.status(204).json({ 'message': `No Payment matching Id ${req.params.id} has been found.` });
+        const anyOrderHeader = await OrderHeader.findOne({ where: { PaymentId: req.params.id } });
+        if(anyOrderHeader)
+            return res.status(204).json({ 'message': 'Payment you tried to delete is used by at least 1 OrderHeader.' });
         await Payment.destroy({
-            where: {
-                id: id
-            }
-        })
+            where: { id: req.params.id } }
+        );
         res.json("Deleted successfully.");
     },
-    // get all Payments
+    
     getAll: async (req, res) => {
         const payments = await Payment.findAll();
+        if (!payments.length) 
+            return res.status(204).json({ 'message': 'No Payments found.' });
         res.json(payments);
     },
-    // get Payment /w specific id
+    
     getById: async (req, res) => {
-        const id = req.params.id
-        const payment = await Payment.findByPk(id);
+        const payment = await Payment.findOne({ where: { id: req.params.id } });
+        if(!payment)
+            return res.status(204).json({ 'message': `No Payment matching Id ${req.params.id} has been found.` });
         res.json(payment);
     },
-    // get Payment by name
+    
     getByName: async (req, res) => {
-        const name = req.params.name
-        const payment = await Payment.findOne({ where: { name: name } });
+        const payment = await Payment.findOne({ where: { name: req.params.name } });
+        if(!payment)
+            return res.status(204).json({ 'message': `No Payment matching Name '${req.params.name}' has been found.` });
         res.json(payment);
     },
 }
